Jenkins file updates

+++ Jenkinsfile
@@ -49,20 +49,27 @@ pipeline{
 		     }
 		}		
 		stage('Deploy'){
-		     when {
-				branch 'Catfish-2.0-calendar-el-block'
-             }
+//		     when {
+//				branch 'Catfish-2.0-calendar-el-block'
+//             }
 		    steps{
-				bat "rm Catfish\\appsettings.json"
-				bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:iisApp="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:iisApp="catfish-dev.artsrn.ualberta.ca" -enableRule:AppOffline """   
+				script{
+					if (env.BRANCH_NAME == 'Catfish-2.0-calendar-el-block'){
+						bat "rm Catfish\\appsettings.json"
+						bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:iisApp="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:iisApp="catfish-dev.artsrn.ualberta.ca" -enableRule:AppOffline """   	
+					}
+					if (env.BRANCH_NAME == 'Catfish-2.0-Interface-Testing'){
+						bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:iisApp="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:iisApp="catfish-test.artsrn.ualberta.ca" -enableRule:AppOffline """   
+					}					
+				}
 		    }
-		    when {
-				branch 'Catfish-2.0-Interface-Testing'
-             }
-		    steps{
-				bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:iisApp="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:iisApp="catfish-test.artsrn.ualberta.ca" -enableRule:AppOffline """   
-				//bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:contentPath="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:contentPath="E:\\inetpub\\wwwroot2\\catfish-test.artsrn.ualberta.ca" -enableRule:AppOffline """   
-		    }
+//		    when {
+//				branch 'Catfish-2.0-Interface-Testing'
+//             }
+//		    steps{
+//				bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:iisApp="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:iisApp="catfish-test.artsrn.ualberta.ca" -enableRule:AppOffline """   
+//				//bat """ "C:\\Program Files\\IIS\\Microsoft Web Deploy V3\\msdeploy.exe"  -verb:sync -source:contentPath="${WORKSPACE}\\Catfish\\bin\\Release\\netcoreapp3.1\\publish" -dest:contentPath="E:\\inetpub\\wwwroot2\\catfish-test.artsrn.ualberta.ca" -enableRule:AppOffline """   
+//		    }
 		}		
 		stage('Test'){
 		     when {
